[#150279546] Architect wants minimal, open-source web browser based XJ Music player, in order to embed XJ Music on any website, and ensure that the experience is as widely accessible as possible.  \   * [ui/player-ui]       - Nginx serves Player Web at the /player/ URI of a Hub instance       - Gulp pipeline distributes javascript, css, images       - initiate project at Github using no framework, and Gulp workflow that is good for a Javascript embeddable app       - The primary system of record for software development at XJ is ***always*** the primary xj http://code.outright.io/xj/xj repo-- we will determine a strategy for publishing downstream releases to github       - Determine some kind of development deployment strategy that forwards localhost to production Hub ***(?)***       - Determine some kind of development deployment strategy that forwards localhost to development Hub ***(?)*** for local development       - Player is embedded along with any number of key-value parameters encoded in the URL       - debug=true enables debug mode, otherwise player writes as little as possible to DOM
[#154043509] Hub UI to instantiate an instance of an XJ Player (ergo its WebAudio) when playback is requested, in order to make optimal use of client-side resources.
 \
  * [hub-ui]
      - Refactored legacy `player-info` component to new `player-embed` component which writes an embedded player to an iframe
      - Chain Segments index now autonomously updates currently playing chain from cached data, without actually communicating with the player about what its current time is, instead inferring this from Hub UI's own record of the time playback began.

[#150279568] Artist expects to be able to open XJ in many tabs.

[#158081635] Developer wants excellent testing experience for XJ Player Web JavaScript application
 \
  * [jest] testing for entire XJ platform front-end JavaScript web UI is done from the project root
  |
  * [bin/ui/build] + [bin/ui/test] Jest testing of player-web is included in main platform build and test

[#158274389] Developer wants to embed Player UI with hash of key-values in URL hash

[#156410528] Deprecate Link in favor of ChainSegment
 \
  * [core/migration/57]
      - RENAME TABLE `link` TO `segment`
      - RENAME TABLE `link_chord` TO `segment_chord`
      - RENAME TABLE `link_meme` TO `segment_meme`
      - RENAME TABLE `link_message` TO `segment_message`
      - RENAME TABLE `chain_pattern` TO `chain_sequence`
      - RENAME TABLE `pattern` TO `sequence`
      - RENAME TABLE `pattern_meme` TO `sequence_meme`
      - RENAME TABLE `phase` TO `pattern`
      - RENAME TABLE `phase_chord` TO `pattern_chord`
      - RENAME TABLE `phase_meme` TO `pattern_meme`
      - RENAME TABLE `phase_event` TO `pattern_event`

[#155730614] Artist wants Sequence and Pattern (formerly Pattern and Phase) to be named according to musical norms, in order to make optimal sense of XJ as a musical instrument.

[#158319467] Developer expects Not to see a "found existing chain with this embed key" error when the only possible suspect is the chain itself.

[#150279540] Unauthenticated public Client wants to access a Chain by embed key (as alias for chain id) in order to provide data for playback.
 \
  * [core] ChainDAO and SegmentDAO have methods to read chain by embed key INSTEAD of access control
  |
  * [hub] Chain and Segments segments have methods to read chain by embed key instead of access control token

[#150279533] Engineer expects when Chain is set to erase state, its `embed_key` is set to null immediately, in order to prevent public consumption of a chain in erase state, and enable a new chain to be created immediately with that `embed_key`

+++ core/src/main/java/io/xj/core/cache/entity/impl/EntityClassCacheItem.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2017, Outright Mental Inc. (http://outright.io) All Rights Reserved.
+// Copyright (c) 2018, XJ Music Inc. (https://xj.io) All Rights Reserved.
 package io.xj.core.cache.entity.impl;
 
 import io.xj.core.config.Config;
@@ -32,7 +32,7 @@ public class EntityClassCacheItem<N extends Entity> {
    */
   public Boolean isValid() {
     return TimestampUTC.now().toInstant().getEpochSecond() <
-      createdAt.toInstant().getEpochSecond() + Config.entityCacheSeconds();
+      createdAt.toInstant().getEpochSecond() + (long) Config.entityCacheSeconds();
   }
 
   /**
