+++ core/src/test/java/io/xj/core/dao/AudioEventIT.java
@@ -45,21 +45,21 @@ public class AudioEventIT {
     IntegrationTestEntity.insertUser(2, "john", "john@email.com", "http://pictures.com/john.gif");
     IntegrationTestEntity.insertUserRole(1, 2, UserRoleType.Admin);
 
-    // Library "palm tree" has pattern "leaves" and pattern "coconuts"
+    // Library "palm tree" has sequence "leaves" and sequence "coconuts"
     IntegrationTestEntity.insertLibrary(1, 1, "palm tree");
 
-    // Pattern "leaves" has instruments "808" and "909"
+    // Sequence "leaves" has instruments "808" and "909"
     IntegrationTestEntity.insertInstrument(1, 1, 2, "808 Drums", InstrumentType.Percussive, 0.9);
     IntegrationTestEntity.insertInstrument(2, 1, 2, "909 Drums", InstrumentType.Percussive, 0.8);
 
     // Instrument "808" has Audio "Beat"
-    IntegrationTestEntity.insertAudio(1, 1, "Published", "Beat", "https://static.xj.io/19801735098q47895897895782138975898.wav", 0.01, 2.123, 120.0, 440);
+    IntegrationTestEntity.insertAudio(1, 1, "Published", "Beat", "https://static.xj.io/19801735098q47895897895782138975898.wav", 0.01, 2.123, 120.0, 440.0);
 
     // Audio "Drums" has events "KICK" and "SNARE" 2x each
-    IntegrationTestEntity.insertAudioEvent(1, 1, 2.5, 1, "KICK", "Eb", 0.8, 1.0);
-    IntegrationTestEntity.insertAudioEvent(2, 1, 3, 1, "SNARE", "Ab", 0.1, 0.8);
-    IntegrationTestEntity.insertAudioEvent(3, 1, 0, 1, "KICK", "C", 0.8, 1.0);
-    IntegrationTestEntity.insertAudioEvent(4, 1, 1, 1, "SNARE", "G", 0.1, 0.8);
+    IntegrationTestEntity.insertAudioEvent(1, 1, 2.5, 1.0, "KICK", "Eb", 0.8, 1.0);
+    IntegrationTestEntity.insertAudioEvent(2, 1, 3.0, 1.0, "SNARE", "Ab", 0.1, 0.8);
+    IntegrationTestEntity.insertAudioEvent(3, 1, (double) 0, 1.0, "KICK", "C", 0.8, 1.0);
+    IntegrationTestEntity.insertAudioEvent(4, 1, 1.0, 1.0, "SNARE", "G", 0.1, 0.8);
 
     // Instantiate the test subject
     testDAO = injector.getInstance(AudioEventDAO.class);
@@ -83,7 +83,7 @@ public class AudioEventIT {
       .setPosition(0.42)
       .setTonality(0.92)
       .setVelocity(0.72)
-      .setAudioId(BigInteger.valueOf(1));
+      .setAudioId(BigInteger.valueOf(1L));
 
     JSONObject result = JSON.objectFrom(testDAO.create(access, inputData));
 
@@ -126,7 +126,7 @@ public class AudioEventIT {
       .setPosition(0.0)
       .setTonality(1.0)
       .setVelocity(1.0)
-      .setAudioId(BigInteger.valueOf(2));
+      .setAudioId(BigInteger.valueOf(2L));
 
     testDAO.create(access, inputData);
   }
@@ -138,11 +138,11 @@ public class AudioEventIT {
       "accounts", "1"
     ));
 
-    AudioEvent result = testDAO.readOne(access, BigInteger.valueOf(4));
+    AudioEvent result = testDAO.readOne(access, BigInteger.valueOf(4L));
 
     assertNotNull(result);
-    assertEquals(BigInteger.valueOf(4), result.getId());
-    assertEquals(BigInteger.valueOf(1), result.getAudioId());
+    assertEquals(BigInteger.valueOf(4L), result.getId());
+    assertEquals(BigInteger.valueOf(1L), result.getAudioId());
     assertEquals(Double.valueOf(1.0), result.getDuration());
     assertEquals("SNARE", result.getInflection());
     assertEquals("G", result.getNote());
@@ -158,7 +158,7 @@ public class AudioEventIT {
       "accounts", "326"
     ));
 
-    AudioEvent result = testDAO.readOne(access, BigInteger.valueOf(1));
+    AudioEvent result = testDAO.readOne(access, BigInteger.valueOf(1L));
 
     assertNull(result);
   }
@@ -170,10 +170,10 @@ public class AudioEventIT {
       "accounts", "1"
     ));
 
-    JSONArray result = JSON.arrayOf(testDAO.readAll(access, ImmutableList.of(BigInteger.valueOf(1))));
+    JSONArray result = JSON.arrayOf(testDAO.readAll(access, ImmutableList.of(BigInteger.valueOf(1L))));
 
     assertNotNull(result);
-    assertEquals(4, result.length());
+    assertEquals(4L, (long) result.length());
     JSONObject result1 = (JSONObject) result.get(0);
     assertEquals("KICK", result1.get("inflection"));
     JSONObject result2 = (JSONObject) result.get(1);
@@ -186,10 +186,10 @@ public class AudioEventIT {
 
   @Test
   public void readAllFirstEventsForInstrument() throws Exception {
-    Collection<AudioEvent> result = testDAO.readAllFirstEventsForInstrument(Access.internal(), BigInteger.valueOf(1));
+    Collection<AudioEvent> result = testDAO.readAllFirstEventsForInstrument(Access.internal(), BigInteger.valueOf(1L));
 
     assertNotNull(result);
-    assertEquals(1, result.size());
+    assertEquals(1L, (long) result.size());
     AudioEvent result1 = result.iterator().next();
     assertEquals("KICK", result1.getInflection());
   }
@@ -201,10 +201,10 @@ public class AudioEventIT {
       "accounts", "345"
     ));
 
-    JSONArray result = JSON.arrayOf(testDAO.readAll(access, ImmutableList.of(BigInteger.valueOf(1))));
+    JSONArray result = JSON.arrayOf(testDAO.readAll(access, ImmutableList.of(BigInteger.valueOf(1L))));
 
     assertNotNull(result);
-    assertEquals(0, result.length());
+    assertEquals(0L, (long) result.length());
   }
 
   @Test(expected = BusinessException.class)
@@ -221,7 +221,7 @@ public class AudioEventIT {
       .setTonality(1.0)
       .setVelocity(1.0);
 
-    testDAO.update(access, BigInteger.valueOf(3), inputData);
+    testDAO.update(access, BigInteger.valueOf(3L), inputData);
   }
 
   @Test(expected = BusinessException.class)
@@ -236,9 +236,9 @@ public class AudioEventIT {
       .setPosition(0.0)
       .setTonality(1.0)
       .setVelocity(1.0)
-      .setAudioId(BigInteger.valueOf(2));
+      .setAudioId(BigInteger.valueOf(2L));
 
-    testDAO.update(access, BigInteger.valueOf(2), inputData);
+    testDAO.update(access, BigInteger.valueOf(2L), inputData);
   }
 
   @Test(expected = BusinessException.class)
@@ -254,16 +254,16 @@ public class AudioEventIT {
       .setPosition(0.0)
       .setTonality(1.0)
       .setVelocity(1.0)
-      .setAudioId(BigInteger.valueOf(287));
+      .setAudioId(BigInteger.valueOf(287L));
 
     try {
-      testDAO.update(access, BigInteger.valueOf(3), inputData);
+      testDAO.update(access, BigInteger.valueOf(3L), inputData);
 
     } catch (Exception e) {
-      AudioEvent result = testDAO.readOne(Access.internal(), BigInteger.valueOf(3));
+      AudioEvent result = testDAO.readOne(Access.internal(), BigInteger.valueOf(3L));
       assertNotNull(result);
       assertEquals("KICK", result.getInflection());
-      assertEquals(BigInteger.valueOf(1), result.getAudioId());
+      assertEquals(BigInteger.valueOf(1L), result.getAudioId());
       throw e;
     }
   }
@@ -281,21 +281,21 @@ public class AudioEventIT {
       .setPosition(0.42)
       .setTonality(0.92)
       .setVelocity(0.72)
-      .setAudioId(BigInteger.valueOf(1));
+      .setAudioId(BigInteger.valueOf(1L));
 
-    testDAO.update(access, BigInteger.valueOf(1), inputData);
+    testDAO.update(access, BigInteger.valueOf(1L), inputData);
 
-    AudioEvent result = testDAO.readOne(Access.internal(), BigInteger.valueOf(1));
+    AudioEvent result = testDAO.readOne(Access.internal(), BigInteger.valueOf(1L));
     assertNotNull(result);
     assertEquals("POPPYCOCK", result.getInflection());
     assertEquals((Double) 1.2, result.getDuration());
     assertEquals((Double) 0.42, result.getPosition());
     assertEquals(0.92, result.getTonality(), 0.01);
     assertEquals(0.72, result.getVelocity(), 0.01);
-    assertEquals(BigInteger.valueOf(1), result.getAudioId());
+    assertEquals(BigInteger.valueOf(1L), result.getAudioId());
   }
 
-  // future test: DAO cannot update Pattern to a User or Library not owned by current session
+  // future test: DAO cannot update Sequence to a User or Library not owned by current session
 
   @Test
   public void delete() throws Exception {
@@ -304,9 +304,9 @@ public class AudioEventIT {
       "accounts", "1"
     ));
 
-    testDAO.destroy(access, BigInteger.valueOf(1));
+    testDAO.destroy(access, BigInteger.valueOf(1L));
 
-    AudioEvent result = testDAO.readOne(Access.internal(), BigInteger.valueOf(1));
+    AudioEvent result = testDAO.readOne(Access.internal(), BigInteger.valueOf(1L));
     assertNull(result);
   }
 
@@ -317,7 +317,7 @@ public class AudioEventIT {
       "accounts", "2"
     ));
 
-    testDAO.destroy(access, BigInteger.valueOf(1));
+    testDAO.destroy(access, BigInteger.valueOf(1L));
   }
 
 }
